Fixed cookie values being used as cookie names with empty values
This has been a mortal bug since version 0.1.0.2, which randomises the
IV to improve security, forcing cookie names to not be encrypted so as
to have deterministic names to allow removing/changing cookies. Sadly,
it still tries to decrypt the names as I forgot to amend the decryptor
to compensate for the change to the encryptor.

I never realised it was broken because I was using a vendored version,
since I never got around to using the extracted version on any project
which got that far.

To prevent this from happening again, I've finally added a test suite,
like I've kept meaning to do. Well, now it's done. Sorry to anyone who
spent time trying to figure out why that doomed version wasn't working
for them, or worse, upgraded from the working 0.1.0.0 to 0.1.0.2.

In my defence, the README says don't depend on this for serious stuff,
as I'm not a cryptographer; perhaps I should also mention that I am an
idiot.

Seriously, guys. This is just thrown out there with little thought for
its utility. Caveat emptor. It's small enough that it can be read in a
sitting (excluding dependencies). If you find it useful, by all means,
use it, but as mentioned in the licence, the responsibility is on your
shoulders to verify that it works, that it's safe, etc.

I only publish this on Hackage for my own use.