Fix incorrect use of comma in signed cookie value
According to RFC 6265, this is an invalid character in a cookie value.

https://tools.ietf.org/html/rfc6265#section-4.1.1

Presumably, this is for backward compatibility, since it was used as a
way to separate multiple cookies as per RFC 2109:

  >Informally, the Set-Cookie response header comprises the token Set-
  >Cookie:, followed by a comma-separated list of one or more cookies.

  — https://tools.ietf.org/html/rfc2109#section-4.2.2

The new way to set multiple cookies, according to RFC 6265, is to send
multiple Set-Cookie headers.

Most modern clients will not split cookies with commas into parts, but
some will, treating the second cookie as blank.

Since we were using a comma as a separator between the signature and a
prefix of `signature`, the signature would be treated by those clients
as its own cookie.